requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymysql==1.1.0
python-dotenv==1.0.1
dbutils==3.0.3  # Note: package name is lowercase
